+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.8 (SNAPSHOT)
-* 
+* Add reset, reset-db, and reset-doc targets
 
 ### 0.0.7
 * Inspect invalid item object to assist troubleshooting